1. Corrected histogram method (there was an error of splitting in right boundary) 2. Corrected types of time in bootstrap, boosting (from int32 to float32)

+++ survivors/ensemble/boosting.py
@@ -69,7 +69,7 @@ class BoostingCRAID(BaseEnsemble):
         self.features = X.columns
         X = X.reset_index(drop=True)
         X[cnt.CENS_NAME] = y[cnt.CENS_NAME].astype(np.int32)
-        X[cnt.TIME_NAME] = y[cnt.TIME_NAME].astype(np.int32)
+        X[cnt.TIME_NAME] = y[cnt.TIME_NAME].astype(np.float32)
         
         self.X_train = X
         self.X_train["ind_start"] = self.X_train.index
